+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState.hs
@@ -21,7 +21,6 @@ module Cardano.Ledger.Shelley.LedgerState (
   PulsingRewUpdate (..),
   FutureGenDeleg (..),
   InstantaneousRewards (..),
-  KeyPairs, -- deprecated
   LedgerState (..),
   PState (..),
   PPUPPredFailure,